Bug 1257303 - Cast to intptr_t to avoid C4312 on VS2015; r=tbsaunde
Without this, we get: C4312: 'reinterpret_cast': conversion from
'const LONG' to 'void *' of greater size

MozReview-Commit-ID: 4WXmvVUd5c8

UltraBlame original commit: ae01415f56b22fc573dac4cef502d808afb3c62f

+++ accessible/windows/msaa/AccessibleWrap.cpp
@@ -1525,7 +1525,7 @@ AccessibleWrap::GetXPAccessibleFor(const VARIANT& aVarChild)
   
   
   if (!IsProxy()) {
-    void* uniqueID = reinterpret_cast<void*>(-aVarChild.lVal);
+    void* uniqueID = reinterpret_cast<void*>(intptr_t(-aVarChild.lVal));
 
     DocAccessible* document = Document();
     Accessible* child =
